refactor(docs): refactor react documentation
