+++ packages/angular-demo/package.json
@@ -29,6 +29,7 @@
         "npm-run-all": "^4.1.5",
         "progress-bar-webpack-plugin": "^2.1.0",
         "rxjs": "~6.6.0",
+        "sass-loader": "^12.1.0",
         "tslib": "^2.0.0",
         "zone.js": "~0.10.2"
     },
