--- packages/glyph-components/.storybook/manager-head.html
@@ -1,37 +1,0 @@
-<style>
-    * {
-        font-family: "Roboto", sans-serif!important;
-    }
-    .dark {
-        background: #181818 !important;
-        color: #b3b3b3 !important;
-    }
-    .dark .os-host {
-        background: #181818 !important;
-        color: #b3b3b3;
-    }
-    .sidebar-subheading {
-        letter-spacing: 0!important;
-        font-size: 12px!important;
-    }
-    .sidebar-subheading button:focus {
-        color: black!important;
-        box-shadow: none!important;
-    }
-    button[data-action="collapse-ref"]:focus {
-        color: black!important;
-        box-shadow: none!important;
-    }
-    .sidebar-item {
-        padding-top: 12px!important;
-        padding-bottom: 12px!important;
-    }
-    .sidebar-item[data-selected="true"] {
-        background: #777!important;
-    }
-    .sidebar-item svg {
-        margin-top: 0!important;
-        width: 15px!important;
-        height: 15px!important;
-    }
-</style>
