--- packages/glyph-components/.storybook/preview-head.html
@@ -1,83 +1,0 @@
-<style>
-    * {
-        font-family: 'Roboto', sans-serif;
-    }
-    .dark {
-        background: #181818 !important;
-        color: #b3b3b3 !important;
-    }
-    .docblock-source {
-        background: #f5f5f5 !important;
-        border-left: 3px solid #7979a7 !important;
-        border-radius: 0 !important;
-    }
-    .docblock-source .language-html .token.tag {
-        color: #003f80 !important;
-    }
-    .dark .sbdocs {
-        background: #181818 !important;
-        color: #b3b3b3;
-    }
-    .dark .sidebar-item {
-        color: #bbbbbb !important;
-    }
-    .dark .os-host {
-        background: #181818 !important;
-        color: #b3b3b3;
-    }
-    .dark .sbdocs-pre > div {
-        background: #000 !important;
-        color: #b3b3b3;
-    }
-    .dark tr td {
-        color: white !important;
-    }
-    .dark tr:nth-of-type(2n) {
-        background-color: #3a3a3a !important;
-    }
-    .dark .docblock-code-toggle {
-        background: #0e0e0e !important;
-        border: none !important;
-        color: #b3b3b3;
-    }
-    .dark .docblock-argstable-body tbody tr td {
-        background: #181818 !important;
-    }
-    .sbdocs-blockquote {
-        background: #f5f5f5 !important;
-        padding: 16px !important;
-        border-left: 4px solid #6b6969 !important;
-    }
-    .warning .sbdocs-blockquote {
-        background: #f0f3de !important;
-        padding: 16px !important;
-        border-left: 4px solid #efd03c !important;
-    }
-    .dark .sbdocs-blockquote {
-        background: #313131 !important;
-        padding: 16px !important;
-        border-left: 4px solid #6b6969 !important;
-    }
-    .dark .sbdocs-blockquote .sbdocs {
-        background: transparent !important;
-    }
-    .sbdocs-blockquote * {
-        font-size: 11px !important;
-    }
-    .sbdocs-h2 {
-        border-bottom: none !important;
-    }
-    .sbdocs-h1 {
-        margin-bottom: 24px !important;
-    }
-    .sbdocs-h6 {
-        font-weight: 400 !important;
-        margin-bottom: 10px !important;
-    }
-    .properties h1 {
-        display: none;
-    }
-    .properties h2 {
-        display: none;
-    }
-</style>
