+++ packages/glyph-components/.storybook/preview.js
@@ -1,6 +1,7 @@
 import { defineCustomElements } from '../dist/esm/loader';
 
 import '!style-loader!css-loader!sass-loader!../dist/glyph-components/glyph-components.css';
+import '!style-loader!css-loader!sass-loader!./../../../.storybook/re-style.scss';
 
 defineCustomElements();
 
@@ -48,42 +49,6 @@ export const parameters = {
         },
     },
 };
-export const globalTypes = {
-    theme: {
-        name: 'Theme',
-        description: 'Global theme for components',
-        defaultValue: 'light',
-        toolbar: {
-            icon: 'edit',
-            title: 'Theme',
-            right: 'Theme',
-            // Array of plain string values or MenuItem shape (see below)
-            items: [ 'light', 'dark' ],
-        },
-    },
-    locale: {
-        name: 'Locale',
-        description: 'Internationalization locale',
-        defaultValue: 'en',
-        toolbar: {
-            icon: 'globe',
-            items: [
-                { value: 'en', right: '🇺🇸', title: 'English' },
-                { value: 'es', right: '🇪🇸', title: 'Español' },
-            ],
-        },
-    },
-    fontSize: {
-        name: 'Font size',
-        description: 'Font size base',
-        defaultValue: 'var(--gui-font--base)',
-        toolbar: {
-            icon: 'expand',
-            items: [
-                { value: '10px', right: '10px', title: 'Small' },
-                { value: 'base', right: '12px', title: 'Base' },
-                { value: '15px', right: '15px', title: 'Big' },
-            ],
-        },
-    },
-};
+
+import { globalTypes } from './../../../.storybook/globals';
+export { globalTypes };
