+++ packages/glyph-components/package.json
@@ -37,6 +37,7 @@
     "npm-run-all": "^4.1.5",
     "numeral": "^2.0.6",
     "perfect-scrollbar": "^1.5.0",
+    "sass-loader": "^12.1.0",
     "sortablejs": "^1.13.0",
     "stencil-click-outside": "^1.7.1"
   },
