+++ packages/glyph-components/src/components/input/input.scss
@@ -8,6 +8,7 @@
 
         input {
             flex: 1;
+            background: transparent;
         }
     }
