+++ packages/glyph-components/src/global/themes/_dark.scss
@@ -5,6 +5,7 @@
     --gui-color--background--lighter: #202020;
 
     --gui-color: #eee;
+    --gui-color--lighter: #505050;
     --gui-color--background--contrast: var(--gui-color);
     --gui-color--contrast: var(--gui-color--background--light);
     --gui-color--modern: #2d2d2d;
