+++ packages/glyph-components/src/stories/components.introduction.stories.mdx
@@ -23,7 +23,13 @@ is the correct project import and pass attributes to components correctly.
 
 ## Install
 
-🚧 The library is not currently published
+```bash
+  # yarn 
+  yarn add glyph-components 
+  
+  # npm 
+  npm i glyph-components
+```
 
 ## Use
