+++ packages/glyph-core/package.json
@@ -22,6 +22,7 @@
         "lodash-es": "^4.17.21",
         "npm-run-all": "^4.1.5",
         "redux": "^4.0.5",
+        "sass-loader": "^12.1.0",
         "typescript": "^4.2.3"
     },
     "devDependencies": {
