+++ packages/glyph-react/.storybook/manager.js
@@ -1,0 +1,1 @@
+import '!style-loader!css-loader!sass-loader!common-storybook/re-style-manager.scss';
