+++ packages/glyph-react/package.json
@@ -24,6 +24,7 @@
         "progress-bar-webpack-plugin": "^2.1.0",
         "prop-types": "^15.7.2",
         "react-scripts": "4.0.3",
+        "sass-loader": "^10",
         "typescript": "^4.1.2",
         "web-vitals": "^1.0.1"
     },
@@ -69,18 +70,20 @@
         "@babel/preset-react": "^7.13.13",
         "@babel/preset-typescript": "^7.13.0",
         "@rollup/plugin-babel": "^5.3.0",
-        "@storybook/addon-actions": "^6.2.1",
-        "@storybook/addon-essentials": "^6.2.1",
-        "@storybook/addon-links": "^6.2.1",
-        "@storybook/react": "^6.2.1",
+        "@storybook/addon-actions": "^6.3.1",
+        "@storybook/addon-essentials": "^6.3.1",
+        "@storybook/addon-links": "^6.3.1",
+        "@storybook/react": "^6.3.1",
         "@types/prop-types": "^15.7.3",
         "babel-loader": "^8.2.2",
+        "common-storybook": "file:../../.storybook",
         "react-redux": "^7.2.3",
         "redux": "^4.0.5",
         "rollup": "^2.44.0",
         "rollup-plugin-copy": "^3.4.0",
         "rollup-plugin-delete": "^2.0.0",
         "rollup-plugin-peer-deps-external": "^2.2.4",
-        "rollup-plugin-typescript2": "^0.30.0"
+        "rollup-plugin-typescript2": "^0.30.0",
+        "sass": "^1.35.1"
     }
 }
