+++ packages/react-demo/package.json
@@ -17,6 +17,7 @@
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.3",
     "react-scripts": "4.0.3",
+    "sass-loader": "^12.1.0",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
   },
