+++ stories/style-guide/colors.stories.mdx
@@ -92,6 +92,9 @@ Dark theme is based on Light color variables with some modifications.
   <ColorItem title="Contrast" subtitle="Contrast color over dark backgrounds" colors={["rgb(49,49,49)"]} />
   <span class="color-palette--third">--gui-color--contrast</span>
 
+  <ColorItem title="Lighter" subtitle="Third texts; Dimished and disabled options" colors={["#505050"]} />
+  <span class="color-palette--third">--gui-color--lighter</span>
+
   <ColorItem title="Modern" subtitle="Special text color for moder interfaces" colors={['#2d2d2d']} />
   <span class="color-palette--third">--gui-color--modern</span>
